refactor into one file per class and move exceptions into own namespace
