+++ src/Microsoft.Identity.Client.Extensions.Msal/Providers/DefaultManagedIdentityConfiguration.cs
@@ -1,0 +1,20 @@
+using Microsoft.Extensions.Configuration;
+
+namespace Microsoft.Identity.Client.Extensions.Msal.Providers
+{
+    internal class DefaultManagedIdentityConfiguration : IManagedIdentityConfiguration
+    {
+        private readonly IConfiguration _config;
+
+        public DefaultManagedIdentityConfiguration(IConfiguration config)
+        {
+            _config = config;
+        }
+
+        public string ManagedIdentitySecret => _config.GetValue<string>(Constants.ManagedIdentitySecretEnvName);
+
+        public string ManagedIdentityEndpoint => _config.GetValue<string>(Constants.ManagedIdentityEndpointEnvName);
+
+        public string ClientId => _config.GetValue<string>(Constants.AzureClientIdEnvName);
+    }
+}
