+++ src/Microsoft.Identity.Client.Extensions.Msal/Providers/ServicePrincipalProvider.cs
@@ -183,146 +183,4 @@ namespace Microsoft.Identity.Client.Extensions.Msal.Providers
             _logger?.Log(level, $"{nameof(ServicePrincipalTokenProvider)}.{memberName} :: {message}");
         }
     }
-
-    /// <summary>
-    /// IManagedIdentityConfiguration provides the configurable properties for the ManagedIdentityProbe
-    /// </summary>
-    internal interface IServicePrincipalConfiguration
-    {
-        /// <summary>
-        /// CertificateBase64 is the base64 encoded representation of an x509 certificate
-        /// </summary>
-        string CertificateBase64 { get; }
-
-        /// <summary>
-        /// CertificateThumbprint is the thumbprint of the certificate in the Windows Certificate Store
-        /// </summary>
-        string CertificateThumbprint { get; }
-
-        /// <summary>
-        /// CertificateSubjectDistinguishedName is the subject distinguished name of the certificate in the Windows Certificate Store
-        /// </summary>
-        string CertificateSubjectDistinguishedName { get; }
-
-        /// <summary>
-        /// CertificateStoreName is the name of the certificate store on Windows where the certificate is stored
-        /// </summary>
-        string CertificateStoreName { get; }
-
-        /// <summary>
-        /// CertificateStoreLocation is the location of the certificate store on Windows where the certificate is stored
-        /// </summary>
-        string CertificateStoreLocation { get; }
-
-        /// <summary>
-        /// TenantId is the AAD TenantID
-        /// </summary>
-        string TenantId { get; }
-
-        /// <summary>
-        /// ClientId is the service principal (application) ID
-        /// </summary>
-        string ClientId { get; }
-
-        /// <summary>
-        /// ClientSecret is the service principal (application) string secret
-        /// </summary>
-        string ClientSecret { get; }
-
-        /// <summary>
-        /// Authority is the URI pointing to the AAD endpoint
-        /// </summary>
-        string Authority { get; }
-    }
-
-    internal class DefaultServicePrincipalConfiguration : IServicePrincipalConfiguration
-    {
-        private readonly IConfiguration _config;
-
-        public DefaultServicePrincipalConfiguration(IConfiguration config)
-        {
-            _config = config;
-        }
-
-        public string ClientId => _config.GetValue<string>(Constants.AzureClientIdEnvName);
-
-        public string CertificateBase64 => _config.GetValue<string>(Constants.AzureCertificateEnvName);
-
-        public string CertificateThumbprint => _config.GetValue<string>(Constants.AzureCertificateThumbprintEnvName);
-
-        public string CertificateStoreName => _config.GetValue<string>(Constants.AzureCertificateStoreEnvName);
-
-        public string TenantId => _config.GetValue<string>(Constants.AzureTenantIdEnvName);
-
-        public string ClientSecret => _config.GetValue<string>(Constants.AzureClientSecretEnvName);
-
-        public string CertificateStoreLocation => _config.GetValue<string>(Constants.AzureCertificateStoreLocationEnvName);
-
-        public string CertificateSubjectDistinguishedName => _config.GetValue<string>(Constants.AzureCertificateSubjectDistinguishedNameEnvName);
-
-        public string Authority => string.IsNullOrWhiteSpace(
-            _config.GetValue<string>(Constants.AadAuthorityEnvName)) ?
-                AadAuthority.DefaultTrustedHost :
-                _config.GetValue<string>(Constants.AadAuthorityEnvName);
-    }
-
-    /// <summary>
-    /// ServicePrincipalTokenProvider fetches an AAD token provided Service Principal credentials.
-    /// </summary>
-    internal class InternalServicePrincipalTokenProvider
-    {
-        private readonly IConfidentialClientApplication _client;
-
-        internal InternalServicePrincipalTokenProvider(string authority, string tenantId, string clientId, string secret, IMsalHttpClientFactory clientFactory = null)
-        {
-            _client = ConfidentialClientApplicationBuilder.Create(clientId)
-                .WithTenantId(tenantId)
-                .WithAuthority(new Uri(authority))
-                .WithClientSecret(secret)
-                .WithHttpClientFactory(clientFactory)
-                .Build();
-        }
-
-        private InternalServicePrincipalTokenProvider(string authority, string tenantId, string clientId, X509Certificate2 cert, IMsalHttpClientFactory clientFactory)
-        {
-            _client = ConfidentialClientApplicationBuilder.Create(clientId)
-                .WithTenantId(tenantId)
-                .WithAuthority(new Uri(authority))
-                .WithCertificate(cert)
-                .WithHttpClientFactory(clientFactory)
-                .Build();
-        }
-
-        /// <summary>
-        ///     ServicePrincipalCredentialProvider constructor to build the provider with a certificate
-        /// </summary>
-        /// <param name="authority">Hostname of the security token service (STS) from which MSAL.NET will acquire the tokens. Ex: login.microsoftonline.com
-        /// </param>
-        /// <param name="tenantId">A string representation for a GUID, which is the ID of the tenant where the account resides</param>
-        /// <param name="clientId">A string representation for a GUID ClientId (application ID) of the application</param>
-        /// <param name="cert">A ClientAssertionCertificate which is the certificate secret for the application</param>
-        public InternalServicePrincipalTokenProvider(string authority, string tenantId, string clientId, X509Certificate2 cert)
-            : this(authority, tenantId, clientId, cert, null)
-        { }
-
-        /// <summary>
-        ///     GetTokenAsync returns a token for a given set of scopes
-        /// </summary>
-        /// <param name="scopes">Scopes requested to access a protected API</param>
-        /// <param name="cancel">Cancellation token to cancel the HTTP token request</param>
-        /// <returns>A token with expiration</returns>
-        public async Task<IToken> GetTokenAsync(IEnumerable<string> scopes, CancellationToken cancel)
-        {
-            var res = await _client.AcquireTokenForClient(scopes)
-                .ExecuteAsync(cancel)
-                .ConfigureAwait(false);
-            return new AccessTokenWithExpiration { ExpiresOn = res.ExpiresOn, AccessToken = res.AccessToken };
-        }
-    }
-
-    internal static class AadAuthority
-    {
-        public const string DefaultTrustedHost = "login.microsoftonline.com";
-        public const string AadCanonicalAuthorityTemplate = "https://{0}/{1}/";
-    }
 }
