+++ tests/Microsoft.Identity.Client.Extensions.Msal.UnitTests/Providers/ChainedTokenProviderTests.cs
@@ -6,6 +6,7 @@ using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Identity.Client.Extensions.Abstractions;
+using Microsoft.Identity.Client.Extensions.Msal.Providers.Exceptions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace Microsoft.Identity.Client.Extensions.Msal.Providers
