+++ tests/Microsoft.Identity.Client.Extensions.Msal.UnitTests/Providers/ExponentialBackoffTest.cs
@@ -3,6 +3,7 @@
 
 using System;
 using System.Threading.Tasks;
+using Microsoft.Identity.Client.Extensions.Msal.Providers.Exceptions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace Microsoft.Identity.Client.Extensions.Msal.Providers
